factor512

+++ mmseg/models/decode_heads/deeppad_head.py
@@ -88,7 +88,7 @@ class DeepPadHead(ASPPHead):
                  **kwargs):
         super(DeepPadHead, self).__init__(**kwargs)
         assert c1_in_channels >= 0
-        self.pad_out_channel = pad_out_channel_factor*c1_channels
+        self.pad_out_channel = int(pad_out_channel_factor*c1_channels)
         self.upsample_f = upsample_factor
         self.dyn_ch = dyn_branch_ch
         self.mask_ch = mask_head_ch
